AI Agent - Azure Codes/Demo3/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
azure-mgmt-logic
python-dotenv